Fix broken tests by overwriting Base functions
Overwrite Base functions to use `ismissing` rather than `=== missing`.
This is enough to fix most broken tests. Also add a few fixes to "normal"
methods to fix the remaining ones.

Keep `==`, `<`, `any` and `all` as they are for now since it does not sound
too useful to keep track of whether all missing values were of the same type
in collections, and it is relatively costly and probably difficult to
get included in Julia Base.